Changes to updating and drives
Fixed drive scanning issues and added more appropriate output.
Setup UF2 file updating.
Fixed minor annoyances.

+++ MainWindow.xaml.cs
@@ -2,7 +2,7 @@
 using System.Windows;
 using Microsoft.Win32;
 using System.Diagnostics;
-using System;
+using Octokit;
 
 namespace DragonInjector_Firmware_Tool
 {
@@ -10,76 +10,18 @@ namespace DragonInjector_Firmware_Tool
     {
         string uf2File;
         string uf2ShortFile;
-        readonly string defaultUF2File = Directory.GetCurrentDirectory() + "\\defaultpayload.uf2";
-        readonly string bootloader = Directory.GetCurrentDirectory() + "\\defaultbootloader.uf2";
+        readonly string defaultUF2File = Directory.GetCurrentDirectory() + "\\payloads\\defaultpayload.uf2";
+        readonly string bootloader = Directory.GetCurrentDirectory() + "\\payloads\\defaultbootloader.uf2";
         
         public MainWindow()
         {
             InitializeComponent();
-
-            DriveInfo[] allDrives = DriveInfo.GetDrives();
-            foreach (DriveInfo d in allDrives)
-            {
-                if (d.IsReady == true)
-                {
-                    if (d.VolumeLabel == "DRAGONBOOT")
-                    {
-                        DriveBox.Items.Add(d.Name);
-                        OutputBox.Content += "\n\\:Found drive: " + (d.Name).Replace(":\\", "");
-                        OutputBox.ScrollToBottom();
-                    }
-                }
-                if (d.IsReady == false)
-                {
-                    //OutputBox.Content += "\n.No drives found";
-                    //OutputBox.ScrollToBottom();
-                }
-            }
-        }
-
-        private void App_DispatcherUnhandledException(object sender, System.Windows.Threading.DispatcherUnhandledExceptionEventArgs e)
-        {
-            Exception theException = e.Exception;
-            string theErrorPath = Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData) + "\\GeneratorTestbedError.txt";
-            using (System.IO.TextWriter theTextWriter = new System.IO.StreamWriter(theErrorPath, true))
-            {
-                DateTime theNow = DateTime.Now;
-                theTextWriter.WriteLine("The error time: " + theNow.ToShortDateString() + " " + theNow.ToShortTimeString());
-                while (theException != null)
-                {
-                    theTextWriter.WriteLine("Exception: " + theException.ToString());
-                    theException = theException.InnerException;
-                }
-            }
-            MessageBox.Show("The program crashed.  A stack trace can be found at:\n" + theErrorPath);
-            e.Handled = true;
-            Application.Current.Shutdown();
+            GetDrives();            
         }
 
         private void DriveButton_Click(object sender, RoutedEventArgs e)
         {
-            OutputBox.Content += "\n...Scanning";
-            int selectedIndex = DriveBox.SelectedIndex;
-            DriveBox.Items.Clear();
-            DriveInfo[] allDrives = DriveInfo.GetDrives();
-            foreach (DriveInfo d in allDrives)
-            {
-                if (d.IsReady == true)
-                {
-                    if (d.VolumeLabel == "DRAGONBOOT")
-                    {
-                        DriveBox.Items.Add(d.Name);
-                        OutputBox.Content += "\n\\:Found drive: " + (d.Name).Replace(":\\", "");
-                        OutputBox.ScrollToBottom();
-                    }
-                }
-                if (d.IsReady == false)
-                {
-                    //OutputBox.Content += "\n.No drives found";
-                    //OutputBox.ScrollToBottom();
-                }
-            }
-            DriveBox.SelectedIndex = selectedIndex;
+            GetDrives();
         }
 
         private void CloseButton_Click(object sender, RoutedEventArgs e)
@@ -89,26 +31,36 @@ namespace DragonInjector_Firmware_Tool
 
         private void CheckUpdateButton_Click(object sender, RoutedEventArgs e)
         {
-            OutputBox.Content += "\n...Checking for updates";
-            OutputBox.ScrollToBottom();
-            LatestBootloaderVersionLabel.Text = "v1.1";
-            LatestFirmwareVersionLabel.Text = "DRAGONBOOT v1.3";
+            GetReleasesAsync();
         }
 
         private void FlashButton_Click(object sender, RoutedEventArgs e)
         {
-            string dest = DriveBox.SelectedItem.ToString() + "\\flash.uf2";
-            if (uf2File != null)
+            if (uf2File != null && DriveBox.Text != "Leave blank for DragonBoot" && DriveBox.SelectedItem != null)
             {
+                string dest = DriveBox.SelectedItem.ToString() + "\\flash.uf2";
                 OutputBox.Content += "\n\\:Copying " + uf2ShortFile + " to " + DriveBox.SelectedItem.ToString().Replace(":\\", "");
                 OutputBox.ScrollToBottom();
                 File.Copy(uf2File, dest, true);
             }
-            else
+            else if (DriveBox.SelectedItem != null && File.Exists(".\\payloads\\defaultpayload.uf2") && uf2File == null || DriveBox.Text == "Leave blank for DragonBoot")
             {
+                OutputBox.Content += "\n...Using default payload";
+                OutputBox.ScrollToBottom();
+                string dest = DriveBox.SelectedItem.ToString() + "\\flash.uf2";
                 OutputBox.Content += "\n\\:Copying default payload to " + DriveBox.SelectedItem.ToString().Replace(":\\", "");
                 OutputBox.ScrollToBottom();
                 File.Copy(defaultUF2File, dest, true);
+            }
+            else if (!File.Exists(".\\payloads\\defaultpayload.uf2"))
+            {
+                OutputBox.Content += "\n!Missing default payload in directory";
+                OutputBox.ScrollToBottom();
+            }
+            else
+            {
+                OutputBox.Content += "\n!No drive selected";
+                OutputBox.ScrollToBottom();
             }
         }
 
@@ -138,11 +90,23 @@ namespace DragonInjector_Firmware_Tool
 
         private void BootloaderButton_Click(object sender, RoutedEventArgs e)
         {
-            string dest = DriveBox.SelectedItem.ToString() + "\\flash.uf2";
-            OutputBox.Content += "\n\\:Updating bootloader on " +  DriveBox.SelectedItem.ToString().Replace(":\\", "");
-            OutputBox.ScrollToBottom();
-            File.Copy(bootloader, dest, true);
-
+            if (DriveBox.SelectedItem != null && File.Exists(".\\payloads\\defaultbootloader.uf2"))
+            {
+                string dest = DriveBox.SelectedItem.ToString() + "\\flash.uf2";
+                OutputBox.Content += "\n\\:Updating bootloader on " + DriveBox.SelectedItem.ToString().Replace(":\\", "");
+                OutputBox.ScrollToBottom();
+                File.Copy(bootloader, dest, true);
+            }
+            else if (!File.Exists(".\\payloads\\defaultbootloader.uf2"))
+            {
+                OutputBox.Content += "\n!Missing default bootloader in directory";
+                OutputBox.ScrollToBottom();
+            }
+            else
+            {
+                OutputBox.Content += "\n!No drive selected";
+                OutputBox.ScrollToBottom();
+            }
         }
 
         private void BootloaderAllButton_Click(object sender, RoutedEventArgs e)
@@ -202,29 +166,74 @@ namespace DragonInjector_Firmware_Tool
         private void PayloadButton_Click(object sender, RoutedEventArgs e)
         {
             PayloadTextBox.Clear();
-            OutputBox.Content += "\n...Using default payload";
-            OutputBox.ScrollToBottom();
         }
 
         private void LogoButton_Click(object sender, RoutedEventArgs e)
         {
             Process.Start("https://www.dragoninjector.com");
         }
+
+        private async System.Threading.Tasks.Task GetReleasesAsync()
+        {
+            OutputBox.Content += "\n...Checking for updates";
+            OutputBox.ScrollToBottom();
+
+            var githubFW = new GitHubClient(new ProductHeaderValue("Nothing"));
+            var releasesFW = await githubFW.Repository.Release.GetAll("dragoninjector-project", "DragonInjector-Firmware");
+            var releaseFW = releasesFW[0];
+            OutputBox.Content += "\n" + "Found firmware release: " + releaseFW.Name.ToString();
+            OutputBox.ScrollToBottom();
+            LatestFirmwareVersionLabel.Text = releaseFW.TagName.ToString();
+            
+            var githubBL = new GitHubClient(new ProductHeaderValue("Nothing"));
+            var releasesBL = await githubBL.Repository.Release.GetAll("dragoninjector-project", "DragonInjector-Bootloader");
+            var releaseBL = releasesBL[0];
+            OutputBox.Content += "\n" + "Found bootloader release: " + releaseBL.Name.ToString();
+            OutputBox.ScrollToBottom();
+            LatestBootloaderVersionLabel.Text = releaseBL.TagName.ToString();
+        }
+        
+        private void GetDrives()
+        {
+            OutputBox.Content += "\n...Scanning for drives";
+            OutputBox.ScrollToBottom();
+            int selectedIndex = DriveBox.SelectedIndex;
+            DriveBox.Items.Clear();
+            DriveInfo[] allDrives = DriveInfo.GetDrives();
+            int badDrive = 0;
+            int goodDrive = 0;
+            foreach (DriveInfo d in allDrives)
+            {
+                if (d.IsReady)
+                {
+                    if (d.VolumeLabel == "DRAGONBOOT")
+                    {
+                        DriveBox.Items.Add(d.Name);
+                        OutputBox.Content += "\n\\:Found drive: " + (d.Name).Replace(":\\", "");
+                        OutputBox.ScrollToBottom();
+                        goodDrive++;
+                    }
+                    else
+                    {
+                        badDrive++;
+                    }
+                }
+            }
+            if (badDrive > 0 && goodDrive == 0)
+            {
+                OutputBox.Content += "\n.No drives found";
+                OutputBox.ScrollToBottom();
+            }
+            DriveBox.SelectedIndex = selectedIndex;
+        }
     }
 }
 
 /*
 TODO:
-
-Really need to fix and add try/catch, especially where drive scan happens
-Check if default files exist first
 Add customization options = show boot logo, show path only, no visual feedback
-Figure out how to pull firmware version
 Add "pressed" states to buttons
-Make check for updates actually, you know, check for updates
-Possibly add more verbose output
-Remove annoying border when drive list is empty
-Add output for no drive(s) selected and fix crash for it
-Add window title
-Fix output when no drives found and device not ready
+Add UF2 version checking
+Create local json and check uf2 version on launch
+Check if default files exist on update and download if they do not
 */
